Refactored tests API, divided helpers into files
